pareto update (#19)
