+++ tpot2/__init__.py
@@ -14,3 +14,4 @@ from .builtin_modules import *
 from .config import *
 from .representations import *
 from .parent_selectors import *
+from .utils import *
